Embed and access user application binary in kernel
Implemented process to build and embed a user-space application (`shell`) into the kernel by converting it to a raw binary (`shell.bin`) and then into an object file (`shell.bin.o`). Added C externs to access the start and end of the binary within the kernel. This enables the kernel to prepare user applications in memory and facilitates switching to userland execution in the future.